Framer refactor (#39)
Update to vFramer

* drawing functions are moved to lib instead of contained only in vFramer. For future development which should allow individual apps to run an inline vFramer for custom drawers.

* vFramer can open a frame-based port which will be read as the background on which to project events.

* vFramer includes tries to read calibration parameters to project the frame-based image onto the events

* border of vITP projected images are now white

* bug fixes in vITP
