+++ lib/include/event-driven/vDraw.h
@@ -26,16 +26,7 @@
 #include <string>
 #include <opencv2/opencv.hpp>
 
-class vDraw;
-
-/**
- * @brief createDrawer returns an instance of a drawer that matches the tag
- * specified
- * @param tag is the code of the drawer
- * @return a pointer to a drawer on success. 0 if no appropiate drawer was
- * found
- */
-vDraw * createDrawer(std::string tag);
+namespace ev {
 
 /**
  * @brief The vDraw class is the base class from which all vDrawers should
@@ -121,9 +112,43 @@ public:
 
 };
 
-class accDraw : public vDraw {
+class isoDraw : public vDraw {
+
+protected:
+
+    //angles
+    double thetaY;
+    double thetaX;
+    double CY, SY;
+    double CX, SX;
+
+    double ts_to_axis;
+    int Zlimit;
+    int imagewidth;
+    int imageheight;
+    int imagexshift;
+    int imageyshift;
+
+    //private functions
+    inline void pttr(int &x, int &y, int &z) {
+        // we want a negative rotation around the y axis (yaw)
+        // a positive rotation around the x axis (pitch) (no roll)
+        // the z should always be negative values.
+        // the points need to be shifted across by negligble amount
+        // the points need to be shifted up by (x = max, y = 0, ts = 0 rotation)
+
+        int xmod = x*CY + z*SY + 0.5; // +0.5 rounds rather than floor
+        int ymod = y*CX - SX*(-x*SY + z*CY) + 0.5;
+        int zmod = y*SX + CX*(-x*SY + z*CY) + 0.5;
+        x = xmod; y = ymod; z = zmod;
+    }
+
+    //image with warped square drawn
+    cv::Mat baseimage;
 
 public:
+
+    void initialise();
 
     static const std::string drawtype;
     virtual void draw(cv::Mat &image, const ev::vQueue &eSet, int vTime);
@@ -133,17 +158,6 @@ public:
 };
 
 class addressDraw : public vDraw {
-
-public:
-
-    static const std::string drawtype;
-    virtual void draw(cv::Mat &image, const ev::vQueue &eSet, int vTime);
-    virtual std::string getDrawType();
-    virtual std::string getEventType();
-
-};
-
-class grayDraw : public vDraw {
 
 public:
 
@@ -209,49 +223,6 @@ public:
 
 };
 
-class clusterDraw : public vDraw {
-
-protected:
-
-    std::map<int, ev::event<ev::GaussianAE>> persistance;
-    int stagnantCount;
-
-public:
-
-    static const std::string drawtype;
-    virtual void draw(cv::Mat &image, const ev::vQueue &eSet, int vTime);
-    virtual std::string getDrawType();
-    virtual std::string getEventType();
-
-};
-
-class circleDraw : public vDraw {
-
-protected:
-
-    std::map<int, ev::event<ev::GaussianAE>> persistance;
-    int stagnantCount;
-
-public:
-
-    static const std::string drawtype;
-    virtual void draw(cv::Mat &image, const ev::vQueue &eSet, int vTime);
-    virtual std::string getDrawType();
-    virtual std::string getEventType();
-
-};
-
-class blobDraw : public vDraw {
-
-public:
-
-    static const std::string drawtype;
-    virtual void draw(cv::Mat &image, const ev::vQueue &eSet, int vTime);
-    virtual std::string getDrawType();
-    virtual std::string getEventType();
-
-};
-
 class interestDraw : public vDraw {
 
 public:
@@ -263,43 +234,14 @@ public:
 
 };
 
-class isoDraw : public vDraw {
+class clusterDraw : public vDraw {
 
 protected:
 
-    //angles
-    double thetaY;
-    double thetaX;
-    double CY, SY;
-    double CX, SX;
-
-    double ts_to_axis;
-    int Zlimit;
-    int imagewidth;
-    int imageheight;
-    int imagexshift;
-    int imageyshift;
-
-    //private functions
-    inline void pttr(int &x, int &y, int &z) {
-        // we want a negative rotation around the y axis (yaw)
-        // a positive rotation around the x axis (pitch) (no roll)
-        // the z should always be negative values.
-        // the points need to be shifted across by negligble amount
-        // the points need to be shifted up by (x = max, y = 0, ts = 0 rotation)
-
-        int xmod = x*CY + z*SY + 0.5; // +0.5 rounds rather than floor
-        int ymod = y*CX - SX*(-x*SY + z*CY) + 0.5;
-        int zmod = y*SX + CX*(-x*SY + z*CY) + 0.5;
-        x = xmod; y = ymod; z = zmod;
-    }
-
-    //image with warped square drawn
-    cv::Mat baseimage;
+    std::map<int, ev::event<ev::GaussianAE>> persistance;
+    int stagnantCount;
 
 public:
-
-    void initialise();
 
     static const std::string drawtype;
     virtual void draw(cv::Mat &image, const ev::vQueue &eSet, int vTime);
@@ -319,38 +261,10 @@ public:
 
 };
 
-class isoCircDraw : public isoDraw {
 
-public:
 
-    static const std::string drawtype;
-    virtual void draw(cv::Mat &image, const ev::vQueue &eSet, int vTime);
-    virtual std::string getDrawType();
-    virtual std::string getEventType();
 
-};
-
-class overlayStereoDraw : public vDraw {
-
-public:
-
-    static const std::string drawtype;
-    virtual void draw(cv::Mat &image, const ev::vQueue &eSet, int vTime);
-    virtual std::string getDrawType();
-    virtual std::string getEventType();
-
-};
-
-class saeDraw : public vDraw {
-
-public:
-
-    static const std::string drawtype;
-    virtual void draw(cv::Mat &image, const ev::vQueue &eSet, int vTime);
-    virtual std::string getDrawType();
-    virtual std::string getEventType();
-
-};
+} //namespace ev::
 
 #endif
