+++ lib/include/event-driven/vIPT.h
@@ -51,7 +51,7 @@ public:
     vIPT();
 
     bool configure(const std::string calibContext, const std::string calibFile);
-    bool showMapProjections();
+    bool showMapProjections(double seconds = 0);
 
     bool sparseForwardTransform(int cam, int &y, int &x);
     bool sparseReverseTransform(int cam, int &y, int &x);
